+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Add `strCount` method
 - Add `sleep` method
 - Add `arrShuffle` method
+- Add `removeSpecialChar` method
 
 ## 0.0.1 (2024-09-02) - Not for Production